backend/requirements.txt
# --- FRAMEWORKS E CORE ---
Django==5.2.8
djangorestframework==3.16.1
asgiref==3.11.0
sqlparse==0.5.4

# --- INFRAESTRUTURA E BANCO DE DADOS (PostgreSQL/MinIO) ---
psycopg2-binary==2.9.11
django-storages==1.14.6
django-minio-storage==0.5.9  # Adicionado para MinIO (abordagem sólida)
boto3==1.42.0
botocore==1.41.6
s3transfer==0.16.0
jmespath==1.0.1
urllib3==2.5.0
requests==2.32.5
charset-normalizer==3.4.4
idna==3.11
certifi==2025.11.12
httplib2==0.31.0
six==1.17.0

# --- AUTENTICAÇÃO E SEGURANÇA ---
djangorestframework-simplejwt==5.5.1  # Adicionado para JWT
python-decouple==3.8

# --- INTELIGÊNCIA ARTIFICIAL (Gemini) ---
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
google-auth==2.43.0
google-auth-httplib2==0.2.1
rsa==4.9.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
uritemplate==4.2.0

# --- UTILIDADES DE DESENVOLVIMENTO E CORE ---
pillow==12.0.0
django-cors-headers==4.9.0
python-dateutil==2.9.0.post0
typing_extensions==4.15.0
# Dependências do Pydantic/Typing (Mantenho para evitar conflitos)
annotated-types==0.7.0
pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
proto-plus==1.26.1
tqdm==4.67.1
cachetools==6.2.2